Add SortingMoves.cs file via upload
All the previous commit are in gitlab. This file includes the sorting class which generates a top 10 best moves in each layer of depth in the alpha beta pruning algorithm giving maximum preference to winning moves, priority to blocking moves and exceptions with a blocking move in the case there is a single square that can block the user from next victory or a winning move to win the game.